+++ src/pagecache/iobuf.rs
@@ -516,7 +516,7 @@ impl IoBufs {
         item: &T,
         header: MessageHeader,
         mut out_buf: &mut [u8],
-        blob_id: Option<Lsn>,
+        heap_reservation: Option<super::heap::Reservation>,
     ) -> Result<()> {
         // we create this double ref to allow scooting
         // the slice forward without doing anything
@@ -527,13 +527,39 @@ impl IoBufs {
             header.serialize_into(out_buf_ref);
         }
 
-        if let Some(blob_id) = blob_id {
+        if let Some(heap_reservation) = heap_reservation {
             // write blob to file
             io_fail!(self, "blob blob write");
-            write_blob(&self.config, header.kind, blob_id, item)?;
+            let mut heap_buf = vec![
+                0;
+                usize::try_from(super::heap::slab_size(
+                    13 + item.serialized_size()
+                ))
+                .unwrap()
+            ];
 
-            let _ = Measure::new(&M.serialize);
-            blob_id.serialize_into(out_buf_ref);
+            let serialization_timer = Measure::new(&M.serialize);
+            heap_buf[0] = header.kind.into();
+            heap_buf[5..13].copy_from_slice(
+                &heap_reservation.heap_id.original_lsn.to_le_bytes(),
+            );
+            let heap_buf_ref: &mut &mut [u8] = &mut &mut heap_buf[13..];
+            item.serialize_into(heap_buf_ref);
+            drop(serialization_timer);
+
+            let mut hasher = crc32fast::Hasher::new();
+            hasher.update(&heap_buf[0..1]);
+            hasher.update(&heap_buf[5..]);
+            let crc = hasher.finalize().to_le_bytes();
+
+            heap_buf[1..5].copy_from_slice(&crc);
+
+            // write the blob pointer and its original lsn into
+            // the log
+            heap_reservation.heap_id.serialize_into(out_buf_ref);
+
+            // write the blob file
+            heap_reservation.complete(&heap_buf)?;
         } else {
             let _ = Measure::new(&M.serialize);
             item.serialize_into(out_buf_ref);