+++ src/pagecache/logger.rs
@@ -2,11 +2,10 @@ use std::fs::File;
 
 use super::{
     arr_to_lsn, arr_to_u32, assert_usize, bump_atomic_lsn, decompress, header,
-    iobuf, lsn_to_arr, pread_exact, pread_exact_or_eof, read_blob, roll_iobuf,
-    u32_to_arr, Arc, BasedBuf, BlobPointer, DiskPtr, IoBuf, IoBufs, LogKind,
-    LogOffset, Lsn, MessageKind, Reservation, Serialize, Snapshot,
-    BATCH_MANIFEST_PID, COUNTER_PID, MAX_MSG_HEADER_LEN, META_PID,
-    MINIMUM_ITEMS_PER_SEGMENT, SEG_HEADER_LEN,
+    iobuf, lsn_to_arr, pread_exact, pread_exact_or_eof, roll_iobuf, u32_to_arr,
+    Arc, BasedBuf, DiskPtr, HeapId, IoBuf, IoBufs, LogKind, LogOffset, Lsn,
+    MessageKind, Reservation, Serialize, Snapshot, BATCH_MANIFEST_PID,
+    COUNTER_PID, MAX_MSG_HEADER_LEN, META_PID, SEG_HEADER_LEN,
 };
 
 use crate::*;
@@ -56,8 +55,9 @@ impl Log {
                 / u64::try_from(self.config.segment_size).unwrap(),
         );
 
+        iobuf::make_durable(&self.iobufs, lsn)?;
+
         if ptr.is_inline() {
-            iobuf::make_durable(&self.iobufs, lsn)?;
             let f = &self.config.file;
             read_message(
                 &**f,
@@ -69,18 +69,19 @@ impl Log {
             // we short-circuit the inline read
             // here because it might not still
             // exist in the inline log.
-            let (_, blob_ptr) = ptr.blob();
-            iobuf::make_durable(&self.iobufs, blob_ptr)?;
-            read_blob(blob_ptr, &self.config).map(|(kind, buf)| {
-                let header = MessageHeader {
-                    kind,
-                    pid,
-                    segment_number: expected_segment_number,
-                    crc32: 0,
-                    len: 0,
-                };
-                LogRead::Blob(header, buf, blob_ptr, 0)
-            })
+            let heap_id = ptr.heap_id().unwrap();
+            self.config.heap.read(heap_id, self.config.use_compression).map(
+                |(kind, buf)| {
+                    let header = MessageHeader {
+                        kind,
+                        pid,
+                        segment_number: expected_segment_number,
+                        crc32: 0,
+                        len: 0,
+                    };
+                    LogRead::Heap(header, buf, heap_id, 0)
+                },
+            )
         }
     }
 
@@ -99,19 +100,20 @@ impl Log {
     }
 
     /// Reserve a replacement buffer for a previously written
-    /// blob write. This ensures the message header has the
-    /// proper blob flag set.
-    pub(super) fn rewrite_blob_pointer(
+    /// heap write. This allows the tiny pointer in the log
+    /// to be migrated to a new segment without copying the
+    /// massive slab in the heap that the pointer references.
+    pub(super) fn rewrite_heap_pointer(
         &self,
         pid: PageId,
-        blob_pointer: BlobPointer,
+        heap_pointer: HeapId,
         guard: &Guard,
     ) -> Result<Reservation<'_>> {
         self.reserve_inner(
             LogKind::Replace,
             pid,
-            &blob_pointer,
-            Some(blob_pointer),
+            &heap_pointer,
+            Some(heap_pointer),
             guard,
         )
     }
@@ -159,7 +161,7 @@ impl Log {
         log_kind: LogKind,
         pid: PageId,
         item: &T,
-        blob_rewrite: Option<Lsn>,
+        heap_rewrite: Option<HeapId>,
         _: &Guard,
     ) -> Result<Reservation<'_>> {
         let _measure = Measure::new(&M.reserve_lat);
@@ -170,14 +172,12 @@ impl Log {
 
         M.reserve_sz.measure(max_buf_len);
 
-        let max_buf_size = (self.config.segment_size
-            / MINIMUM_ITEMS_PER_SEGMENT)
-            - SEG_HEADER_LEN;
+        let max_buf_size = self.config.segment_size - SEG_HEADER_LEN;
 
-        let over_blob_threshold =
+        let over_heap_threshold =
             max_buf_len > u64::try_from(max_buf_size).unwrap();
 
-        assert!(!(over_blob_threshold && blob_rewrite.is_some()));
+        assert!(!(over_heap_threshold && heap_rewrite.is_some()));
 
         let mut printed = false;
         macro_rules! trace_once {
@@ -194,22 +194,22 @@ impl Log {
         let kind = match (
             pid,
             log_kind,
-            over_blob_threshold || blob_rewrite.is_some(),
+            over_heap_threshold || heap_rewrite.is_some(),
         ) {
             (COUNTER_PID, LogKind::Replace, false) => MessageKind::Counter,
-            (META_PID, LogKind::Replace, true) => MessageKind::BlobMeta,
+            (META_PID, LogKind::Replace, true) => MessageKind::HeapMeta,
             (META_PID, LogKind::Replace, false) => MessageKind::InlineMeta,
             (BATCH_MANIFEST_PID, LogKind::Skip, false) => {
                 MessageKind::BatchManifest
             }
             (_, LogKind::Free, false) => MessageKind::Free,
-            (_, LogKind::Replace, true) => MessageKind::BlobNode,
+            (_, LogKind::Replace, true) => MessageKind::HeapNode,
             (_, LogKind::Replace, false) => MessageKind::InlineNode,
-            (_, LogKind::Link, true) => MessageKind::BlobLink,
+            (_, LogKind::Link, true) => MessageKind::HeapLink,
             (_, LogKind::Link, false) => MessageKind::InlineLink,
             other => unreachable!(
                 "unexpected combination of PageId, \
-                 LogKind, and blob status: {:?}",
+                 LogKind, and heap status: {:?}",
                 other
             ),
         };
@@ -259,17 +259,18 @@ impl Log {
                         / u64::try_from(self.config.segment_size).unwrap(),
                 ),
                 pid,
-                len: if over_blob_threshold {
-                    reservation_lsn.serialized_size()
+                len: if over_heap_threshold {
+                    // a HeapId is always 16 bytes
+                    16
                 } else {
                     serialized_len
                 },
             };
 
-            let inline_buf_len = if over_blob_threshold {
+            let inline_buf_len = if over_heap_threshold {
                 usize::try_from(
-                    message_header.serialized_size()
-                        + reservation_lsn.serialized_size(),
+                    // a HeapId is always 16 bytes
+                    message_header.serialized_size() + 16,
                 )
                 .unwrap()
             } else {
@@ -279,7 +280,12 @@ impl Log {
                 .unwrap()
             };
 
-            trace!("reserving buf of len {}", inline_buf_len);
+            trace!(
+                "reserving buf of len {} for pid {} with kind {:?}",
+                inline_buf_len,
+                pid,
+                kind
+            );
 
             // try to claim space
             let prospective_size = buf_offset + inline_buf_len;
@@ -363,22 +369,30 @@ impl Log {
                 reservation_lsn + inline_buf_len as Lsn - 1,
             );
 
-            let blob_id =
-                if over_blob_threshold { Some(reservation_lsn) } else { None };
+            let (heap_reservation, heap_id) = if over_heap_threshold {
+                let heap_reservation = self
+                    .config
+                    .heap
+                    .reserve(serialized_len + 13, reservation_lsn);
+                let heap_id = heap_reservation.heap_id;
+                (Some(heap_reservation), Some(heap_id))
+            } else {
+                (None, None)
+            };
 
             self.iobufs.encapsulate(
                 item,
                 message_header,
                 destination,
-                blob_id,
+                heap_reservation,
             )?;
 
             M.log_reservation_success();
 
-            let pointer = if let Some(blob_id) = blob_id {
-                DiskPtr::new_blob(reservation_lid, blob_id)
-            } else if let Some(blob_rewrite) = blob_rewrite {
-                DiskPtr::new_blob(reservation_lid, blob_rewrite)
+            let pointer = if let Some(heap_id) = heap_id {
+                DiskPtr::new_heap_item(reservation_lid, heap_id)
+            } else if let Some(heap_id) = heap_rewrite {
+                DiskPtr::new_heap_item(reservation_lid, heap_id)
             } else {
                 DiskPtr::new_inline(reservation_lid)
             };
@@ -390,7 +404,7 @@ impl Log {
                 flushed: false,
                 lsn: reservation_lsn,
                 pointer,
-                is_blob_rewrite: blob_rewrite.is_some(),
+                is_heap_item_rewrite: heap_rewrite.is_some(),
                 header_len: usize::try_from(message_header.serialized_size())
                     .unwrap(),
             });
@@ -513,25 +527,25 @@ pub struct SegmentHeader {
 pub enum LogRead {
     /// Successful read, entirely on-log
     Inline(MessageHeader, Vec<u8>, u32),
-    /// Successful read, spilled to its own blob file
-    Blob(MessageHeader, Vec<u8>, BlobPointer, u32),
+    /// Successful read, spilled to a slot in the heap
+    Heap(MessageHeader, Vec<u8>, HeapId, u32),
     /// A cancelled message was encountered
     Canceled(u32),
     /// A padding message used to show that a segment was filled
     Cap(SegmentNumber),
     /// This log message was not readable due to corruption
     Corrupted,
-    /// This blob file is no longer available
-    DanglingBlob(MessageHeader, BlobPointer, u32),
+    /// This heap slot has been replaced
+    DanglingHeap(MessageHeader, HeapId, u32),
     /// This data may only be read if at least this future location is stable
     BatchManifest(Lsn, u32),
 }
 
 impl LogRead {
-    /// Return true if we read a successful Inline or Blob value.
+    /// Return true if we read a successful Inline or Heap value.
     pub fn is_successful(&self) -> bool {
         match *self {
-            LogRead::Inline(..) | LogRead::Blob(..) => true,
+            LogRead::Inline(..) | LogRead::Heap(..) => true,
             _ => false,
         }
     }
@@ -539,7 +553,7 @@ impl LogRead {
     /// Return the underlying data read from a log read, if successful.
     pub fn into_data(self) -> Option<Vec<u8>> {
         match self {
-            LogRead::Blob(_, buf, _, _) | LogRead::Inline(_, buf, _) => {
+            LogRead::Heap(_, buf, _, _) | LogRead::Inline(_, buf, _) => {
                 Some(buf)
             }
             _ => None,
@@ -710,7 +724,7 @@ pub(crate) fn read_message<R: ReadAt>(
     file: &R,
     lid: LogOffset,
     expected_segment_number: SegmentNumber,
-    config: &Config,
+    config: &RunningConfig,
 ) -> Result<LogRead> {
     let _measure = Measure::new(&M.read);
     let segment_len = config.segment_size;
@@ -798,34 +812,26 @@ pub(crate) fn read_message<R: ReadAt>(
             trace!("read pad in segment number {:?}", header.segment_number);
             Ok(LogRead::Cap(header.segment_number))
         }
-        MessageKind::BlobLink
-        | MessageKind::BlobNode
-        | MessageKind::BlobMeta => {
-            let id = arr_to_lsn(&buf);
+        MessageKind::HeapLink
+        | MessageKind::HeapNode
+        | MessageKind::HeapMeta => {
+            assert_eq!(buf.len(), 16);
+            let heap_id = HeapId::deserialize(&mut &buf[..]).unwrap();
 
-            match read_blob(id, config) {
+            match config.heap.read(heap_id, config.use_compression) {
                 Ok((kind, buf)) => {
                     assert_eq!(header.kind, kind);
                     trace!(
-                        "read a successful blob message for blob {} in segment number {:?}",
-                        id,
+                        "read a successful heap message for heap {:?} in segment number {:?}",
+                        heap_id,
                         header.segment_number,
                     );
 
-                    Ok(LogRead::Blob(header, buf, id, inline_len))
+                    Ok(LogRead::Heap(header, buf, heap_id, inline_len))
                 }
-                Err(Error::Io(ref e))
-                    if e.kind() == std::io::ErrorKind::NotFound =>
-                {
-                    debug!(
-                        "underlying blob file not found for blob {} in segment number {:?}",
-                        id, header.segment_number,
-                    );
-                    Ok(LogRead::DanglingBlob(header, id, inline_len))
-                }
-                Err(other_e) => {
-                    debug!("failed to read blob: {:?}", other_e);
-                    Err(other_e)
+                Err(e) => {
+                    debug!("failed to read heap: {:?}", e);
+                    Ok(LogRead::DanglingHeap(header, heap_id, inline_len))
                 }
             }
         }